integrated version: free_buffer_tracker+memory_bank.
the width of 'vc_read_from' and 'vc_written_into' are modified, each wire dedicates to a specified vc ID.

the testbench is also modified, but still some work to do to make the functionality of this dynamically allocated VC works well.

the 'make run' is also modified to adopt this integration.